Use f_bavail instead of f_bfree to compute fs space
bfree doesn't mean what I assumed it means. The actual amount of file
system space available to use for regular processes is bavail, while
bfree describes the "real" amount of free space on the actual file
system.

On ext4, for example, ~5% of blocks are reserved for internal purposes.
This space isn't available to the user during normal operation, and it
doesn't make sense to include it in the free space calculation.